Cleaning up the code, removing unnecesarry bits
