+++ client/app/page.jsx
@@ -8,8 +8,8 @@ import FeaturedDestinations from '../components/FeaturedDestinations/FeaturedDes
 import TopTour from '../components/TopTour/TopTour';
 import ExploreWorld from '../components/Hotel/ExploreWorld';
 import TrendingCites from '../components/Hotel/TrendingCites';
-import Login from '../components/Signig/Login';
-import Signup from '../components/Signig/Signup';
+import Login from '../components/SignIn/Login';
+import Signup from '../components/SignIn/Signup';
 
 export default function Home() {
   return (
